refactor: alterar funcoes de auth para controllers
